+++ client/src/app/components/Background.tsx
@@ -1,5 +1,24 @@
+import { createStyles } from "@mantine/core";
 import React from "react";
-import { useStyles } from "../hooks/useStyles";
+
+const useStyles = createStyles((_theme) => ({
+  container: {
+    zIndex: -100,
+    overflow: "hidden",
+    position: "fixed",
+    height: "100vh",
+    width: "100vw",
+  },
+  circle: {
+    position: "relative",
+    clipPath: "circle(45vw)",
+    top: "0vw",
+    left: "-4vw",
+    height: "134vh",
+    width: "100vw",
+    background: "#9ee6f5",
+  },
+}));
 
 type Props = {};
