+++ client/src/app/components/CodeInput.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import React from "react";
-import { CodeInput } from "./CodeInput";
+import CodeInput from "./CodeInput";
 import CodeInputProvider from "./providers/CodeInputProvider";
 
 describe("<CodeInput />", () => {
