+++ client/src/stories/AppTemplate.stories.tsx
@@ -2,7 +2,8 @@ import { Text } from "@mantine/core";
 import { ComponentMeta, ComponentStory } from "@storybook/react";
 import React from "react";
 import { MemoryRouter } from "react-router-dom";
-import AppTemplate, { Content } from "../app/components/AppTemplate";
+import AppTemplate from "../app/components/AppTemplate";
+import Content from "../app/components/Content";
 
 export default {
   title: "AppTemplate",
